Move all items in project folder to root directory
The .gitignore uses a leading "/" at the beginning, but all of the
matches it is supposed to make are contained in a folder called
Hack-a-Frac. Additionally, it's unusual to have the project contained
within another folder in the repository. This commit moves all of the
project files into the root to solve these two problems.